Update order of images to match instructions
In the instructions wa4e.com & Coursera the 2 images, described as "1st for the assignment" and "2nd in the peer-graded", are shown as 1) pma with date 2) php.ini. The assignment submission page is the opposite, 1) php.ini and 2) pma. Switch the 2 titles around to match instructions.  A few students have commented on this.